feat(ui): make structured plan output denser
