+++ ui/src/domain/Jobs/Details.tsx
@@ -21,6 +21,7 @@ import { useAbortController, usePolling } from "../../hooks";
 import { Job, JobStep } from "../types";
 import { TerminalOutput } from "./TerminalOutput";
 import { getJobOutputRequestUrl, getPublicApiOrigin, isTerrakubeApiUrl } from "./outputUrl";
+import { shouldStepBeExpandedByDefault } from "./stepExpansion";
 import { StructuredPlanOutput } from "./StructuredPlanOutput";
 import { StructuredPlanOutputByStep, normalizeStructuredPlanOutput, normalizeUITemplates } from "./structuredPlan";
 
@@ -29,7 +30,6 @@ type Props = {
 };
 
 const TERMINAL_JOB_STATUSES = new Set(["completed", "noChanges", "failed", "cancelled", "rejected", "notExecuted"]);
-const CLOSED_BY_DEFAULT_STEP_NAMES = new Set(["manual approval pending", "apply pending"]);
 
 export const DetailsJob = ({ jobId }: Props) => {
   const organizationId = sessionStorage.getItem(ORGANIZATION_ARCHIVE);
@@ -59,17 +59,6 @@ export const DetailsJob = ({ jobId }: Props) => {
     }
 
     return TERMINAL_JOB_STATUSES.has(status);
-  };
-
-  const shouldStepBeExpandedByDefault = (item: JobStep) => {
-    const normalizedName = item.name.trim().toLowerCase();
-    const isApprovalPendingStep = CLOSED_BY_DEFAULT_STEP_NAMES.has(normalizedName);
-
-    if (!isApprovalPendingStep) {
-      return true;
-    }
-
-    return item.status !== "waitingApproval" && item.status !== "notExecuted" && item.status !== "pending";
   };
 
   const outputLog = async (output: string | undefined, status: string, signal: AbortSignal) => {
